Функция вынесена в отдельный файл, тест Form исправлен, написан тест для функции фильтрации
